Add two more options to control the output
With --bender-no-sorted-overview, the new section showing
the sorted tests by context and by order within a context
is simply omitted.

With --bender-verbose, the compact progress is replaced
with the normal detailed output, with an eye on long running
test suites where it is important to see the failure as soon
as they happen. In this spirit, we also directly show the
details of the failure.