Simplify report generation by having more stateful Result.
Instead of recomputing the State corresponding to one Result,
we just make the Result#state accessor public.

We also store the "execution count" in the Result.
This number corresponds to the number of results
of the same state created so far.

For example, the first Failing result has 1, second
Failing result has 2, then first Raising result has 1,
and so on.