Avoid secondary error while producing the TEST=... info for a re-run.
The "normal" use case works fine: when you run your test suite through
`rake test` and the usual Rake::TestTask, the file paths of the test
files end up being absolute paths.

When you run the test directly, with:

    ruby spec/my_spec.rb

the path ends up being a relative one and this triggers the following
error:

    .../minitest-bender/lib/minitest-bender/utils.rb:4:in `relative_path': undefined method `gsub' for nil:NilClass (NoMethodError)

This is because the regexp trying to identify the sources in the
'/spec/' or '/test/' directories fails for a 'spec/...' input.

We fix this and also cope with the case of going directly in the
directory containing the test source and running it from there
(e.g. cd spec; ruby my_spec.rb).

If the path to the test source really can't be identified, we say
so instead of failing with a secondary error.