Sorted overview highlights differences from one header to the next.
Precisely because the results are sorted, there is a good deal of redundancy
from one result header to the next. We don't remove that redundancy,
but at least we provide some visual hints about what differs
from one result to the next.
This makes it easier to grasp the structure of the tests.